+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Create geocoding functions `geocode_table_communes()`, `geocode_cities` and `geocode_cities_cp()`
 
+- Create distance computation function `calc_distance()`
+
 # evavelo 0.3.0
 
 - Add button in the UI to download full file and logs